Fix it

+++ src/vsjson.cc
@@ -42,7 +42,7 @@ const char *_vsjson_set_token (vsjson_t *self, const char *ptr, size_t len)
     if (!ptr || !self) return NULL;
 
     if (!len) len = strlen (ptr);
-    if (self->tokensize > len + 1) {
+    if (self->tokensize > (int) len + 1) {
         // fits in
         strncpy (self->token, ptr, len);
         self->token[len] = 0;
@@ -509,7 +509,7 @@ char *vsjson_encode_nstring (const char *string, size_t len)
     if (!encoded) return NULL;
     memset (encoded, 0, capacity);
     encoded[0] = '"';
-    while (*p && p - string < len) {
+    while (*p && p - string < (int) len) {
         switch (*p) {
         case '"':
         case '\\':
